runner/switchboard/requirements.txt
ipython==8.5.0
ipdb
SQLAlchemy==1.4.42
python-dotenv==0.20.0
psycopg2-binary==2.9.5